# openapi/catalog-api.yaml
openapi: 3.1.0
info:
  title: Agentic Commerce Catalog API
  version: 0.1.0
  description: >-
    Technology-neutral HTTP contract for Catalog. The .NET service implements
    GET /health, POST /products, and GET /products/{id}; the remaining Product
    and Category operations are preserved for later vertical slices.
  license:
    name: UNLICENSED
security: []
servers:
  - url: /
    description: Current Catalog API host.
tags:
  - name: Health
    description: Runtime health information.
  - name: Products
    description: Product lifecycle operations.
  - name: Categories
    description: Product category operations.
paths:
  /health:
    get:
      operationId: getHealth
      summary: Check service health
      tags: [Health]
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: The service is healthy.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
              example:
                status: healthy
                service: catalog-api
                version: 0.1.0
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products:
    post:
      operationId: createProduct
      summary: Create a product
      tags: [Products]
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
            example:
              sku: BOOK-001
              name: Domain-Driven Design
              description: A software design reference.
              price: 42.5
      responses:
        "201":
          description: Product created.
          headers:
            Location:
              description: Relative URL of the created Product resource.
              schema:
                type: string
              example: /products/PRODUCT-136fc61e-510d-4d7a-ad63-02e134e1b77f
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEnvelope"
              example:
                data:
                  id: PRODUCT-136fc61e-510d-4d7a-ad63-02e134e1b77f
                  sku: BOOK-001
                  name: Domain-Driven Design
                  description: A software design reference.
                  price: 42.5
                  isActive: true
                  createdAt: "2026-08-06T12:00:00Z"
                  updatedAt: "2026-08-06T12:00:00Z"
                correlationId: 1701dda0-c828-4307-8f4c-11212a8e0412
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      operationId: listProducts
      summary: List products
      tags: [Products]
      x-implementation-status: pending-dotnet
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Products ordered by creation time, newest first.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListEnvelope"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/{id}:
    get:
      operationId: getProduct
      summary: Get a product by ID
      tags: [Products]
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/ProductId"
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Product found.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEnvelope"
              example:
                data:
                  id: PRODUCT-136fc61e-510d-4d7a-ad63-02e134e1b77f
                  sku: BOOK-001
                  name: Domain-Driven Design
                  description: A software design reference.
                  price: 42.5
                  isActive: true
                  createdAt: "2026-08-06T12:00:00Z"
                  updatedAt: "2026-08-06T12:00:00Z"
                correlationId: 10bdd36c-973a-42fe-98fd-113a670f07ec
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/{id}/status:
    patch:
      operationId: updateProductStatus
      summary: Activate or deactivate a product
      tags: [Products]
      x-implementation-status: pending-dotnet
      parameters:
        - $ref: "#/components/parameters/ProductId"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductStatusRequest"
            examples:
              deactivate:
                value:
                  isActive: false
              activate:
                value:
                  isActive: true
      responses:
        "200":
          description: Product status updated.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /categories:
    post:
      operationId: createCategory
      summary: Create a category
      tags: [Categories]
      x-implementation-status: pending-dotnet
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
            example:
              name: Books
              description: Printed and digital books.
      responses:
        "201":
          description: Category created.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      operationId: listCategories
      summary: List categories
      tags: [Categories]
      x-implementation-status: pending-dotnet
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Categories ordered by normalized name.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListEnvelope"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      description: Stable product identifier.
      schema:
        type: string
        pattern: "^PRODUCT-[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
      example: PRODUCT-136fc61e-510d-4d7a-ad63-02e134e1b77f
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: >-
        Caller-provided correlation ID. When omitted or invalid, the service
        generates one and always returns it in the response.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
  headers:
    CorrelationId:
      description: Correlation ID associated with the request.
      schema:
        type: string
  responses:
    BadRequest:
      description: The request is invalid.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            requestSyntax:
              summary: Malformed JSON or incompatible JSON type
              value:
                error:
                  code: REQUEST_VALIDATION_FAILED
                  message: The request payload is invalid.
                  details: {}
                correlationId: 1701dda0-c828-4307-8f4c-11212a8e0412
            productValidation:
              summary: Product business validation failed
              value:
                error:
                  code: PRODUCT_VALIDATION_FAILED
                  message: The product request is invalid.
                  details: {}
                correlationId: 1701dda0-c828-4307-8f4c-11212a8e0412
    NotFound:
      description: The requested resource does not exist.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: PRODUCT_NOT_FOUND
              message: The requested product does not exist.
              details:
                productId: PRODUCT-00000000-0000-4000-8000-000000000099
            correlationId: 10bdd36c-973a-42fe-98fd-113a670f07ec
    Conflict:
      description: A unique catalog value already exists.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: PRODUCT_SKU_ALREADY_EXISTS
              message: A product with this SKU already exists.
              details:
                sku: BOOK-001
            correlationId: 13f789c9-a032-48c6-b1eb-4439d1464657
    InternalServerError:
      description: An unexpected server error occurred.
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: UNEXPECTED_ERROR
              message: An unexpected error occurred.
              details: {}
            correlationId: 7ecf5fea-bd97-42bf-a715-1553798a5180
  schemas:
    Health:
      type: object
      additionalProperties: false
      required: [status, service, version]
      properties:
        status:
          type: string
          const: healthy
        service:
          type: string
          const: catalog-api
        version:
          type: string
          const: 0.1.0
    Product:
      type: object
      additionalProperties: false
      required: [id, sku, name, price, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: "^PRODUCT-[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
        sku:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        price:
          type: number
          minimum: 0
          maximum: 9999999999999999.99
          multipleOf: 0.01
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Category:
      type: object
      additionalProperties: false
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      additionalProperties: false
      required: [sku, name, price]
      properties:
        sku:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        price:
          type: number
          minimum: 0
          maximum: 9999999999999999.99
          multipleOf: 0.01
    UpdateProductStatusRequest:
      type: object
      additionalProperties: false
      required: [isActive]
      properties:
        isActive:
          type: boolean
    CreateCategoryRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
    Error:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ProductEnvelope:
      type: object
      additionalProperties: false
      required: [data, correlationId]
      properties:
        data:
          $ref: "#/components/schemas/Product"
        correlationId:
          type: string
    ProductListEnvelope:
      type: object
      additionalProperties: false
      required: [data, correlationId]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        correlationId:
          type: string
    CategoryEnvelope:
      type: object
      additionalProperties: false
      required: [data, correlationId]
      properties:
        data:
          $ref: "#/components/schemas/Category"
        correlationId:
          type: string
    CategoryListEnvelope:
      type: object
      additionalProperties: false
      required: [data, correlationId]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        correlationId:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, correlationId]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        correlationId:
          type: string